Make moneyFormat more dumb.  Do the infinity checking before calling moneyFormat.
Expand the color set, add reference to Cynthia Brewer's color
configurator (and source of the colors).

And now, the meat: calculate the labels:
  First, -Infinity and Infinity to the ends of the domain, and then
create pairs out of the array.  From here, calculate what the label
would be, for each pair.  Then, add a "No Limit" color (accommodate the
expanded color set).

Now, call the legend's .label() API with the array of labels.