Make -Infinity and Infinity the new extremes of the bins.  Thus, the legend and the viz share the same domain and color range.
Don't filter the data, because now, states without a value for that
particular field are assumed to have "No Limit" to their maximum
contributions.

Adjust the tooltip to handle the empty ones.  And adjust the fill as
well.

Get rid of the intermediary pairs variable, because the Inifinity extremes are statically defined as part of the shared domain array.