Since all the question/answer nodes are being populated via user configuration and Jekyll, we don't need to bother with appending them here.  We just need to respond to them appropriately.
thus, the initSection() function now takes the section name as its first
argument, which helps it choose the right container for its forms.

Since the data behind these nodes is no longer dynamic, we just send the
value from the dropdown as an argument to updateSelectedField(), and it
can query the data object directly with that information.