Move the updatedSelectField() function call outside of the generic init function for now.  Create a legend type for each legend specified in site.data.sections.disclosure.legends.
Add a custom "change" event to the dropdown, which is now populated
statically, with all the data as part of each option. Here, we'll use
d3.select().datum() to capture those HTML data- attributes, and then use
that captured information to change the legend we display, and the
contents of the description field.

This is major progress.