Create an object to hold all the tabulus objects.  Initialize them from an array that Jekyll builds, based on user input about the tabs themselves.
Next, the colorScales (in some cases, multiple) for each tab are
generated as d3 scales with domain and range -- via Jekyll.

Now, it's the jekyll code (the inner loop) repeating itself, but that'll
do for now, at least.

In all cases (except contribution-limits) call the new tabulus() object
to wire up the appropriate tabs.  So at least the d3/js layer is DRYer.