Now, the tabs have memory of the previous query.  So, if you change the dropdowns, then switch away from the tab and then back to the tab, it ill go back to what you last saw.
Simplify, simplify, simplify all the logics.

This time, check for a current question.  If there is one, and the
incoming query is the same, do nothing.  Otherwise, set the active
question to the new query question (or take the current, or set it to
blank).
If we find multiple dropdowns in the container, then let's process the
question and break it down into donor/recipient/branch, and throw them
into the active query.

Now, we cycle through all the dropdowns in the container, and look up
its name in the active query.  If we find a value, set it, and issue the
change signal.