Turn this into @curran style reusable object, to be consistent with grid. Add a few new API (getter/setter) methods:
 1. my.svg()
 2. my.geo()
 4. my.query()

The geo() api call takes the topojson file as argument, and runs some
topojson stuff on it (that used to be the geogrify() function before).

The svg gets initialized as soon as it is sent in.  Thus, the main
function object is empty, but that seems to be ok.

The query method retrieves a query, and if there is a year specified,
then it runs the update() helper function (new).

We don't have to shape the data here, we just accept what is passed in
(add a return my after that).

The update function, then, looks at the query object for how to find the
appropriate value to send into the colorScale (also passed as part of
the query object).

Addeded keyColumn as beach-head to port @curran's logic over from grid.

Added the identity() helper here also.