A number of changes here, in response to the atlas and grid changes. First of all, we now explicitly capture the output of ingest, which gives a nested by year, then by state, arrangement to the data.
The keys, then, feed the year chooser's .data() (via a new helper
function introduced to do just that).

The nested-nested data as is feeds into atlas, which uses the year and
state fields explicitly in order to address the data and get to the
question.

Thus, all the calls to atlas are now similar to grid, except for the
change step, which sends the query object containing all necessary
information at once.

Thus, also, the corpus and atlas initialization methods are gone, and
those statemetnts are in the main function, and show the congruence
(mostly) for the two, which is nice.